server/requirements.txt
flask==3.0.0
flask-socketio==5.3.5
flask-cors==4.0.0
python-dotenv==1.0.0
openai==1.12.0
python-engineio==4.8.0
python-socketio==5.10.0
gevent==23.9.1
gevent-websocket==0.10.1
orjson==3.9.10